fastapi/requirements.txt
# 👉 FastAPI 관련 패키지
fastapi
uvicorn
requests
python-multipart
email-validator>=2.0
python-dotenv
# 👉 AWS S3 연동
boto3

# openai 연동
openai

#db 연동
psycopg2-binary